+++ negotiation/src/main/java/it/eng/negotiation/service/AutomaticNegotiationService.java
@@ -10,6 +10,7 @@ import it.eng.tools.event.AuditEventType;
 import it.eng.tools.model.IConstants;
 import it.eng.tools.service.AuditEventPublisher;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.stereotype.Service;
 
@@ -52,7 +53,7 @@ public class AutomaticNegotiationService {
                                        ContractNegotiationProperties properties,
                                        ContractNegotiationRepository repository,
                                        AuditEventPublisher publisher,
-                                       TaskScheduler taskScheduler) {
+                                       @Qualifier("negotiationTaskScheduler") TaskScheduler taskScheduler) {
         this.apiService = apiService;
         this.properties = properties;
         this.repository = repository;